travis: remove golint errors about unexported types
The operator.Option interface is implemented by types returned by the
various operators option setting function (ConstantProbability, etc.)
The types of those returned objects is unexported on purpose, it is an
implementation detail that is not important to the framework user.
For that reason, the `golint` warnings about unexported types are
excluded in travis configuration file.